made mock values available
