+++ src/components/graphs/AvarageGrade/QsAvarageGradeGraph.tsx
@@ -1,15 +1,15 @@
 import { useEffect, useState } from "react";
 
 import { GraphObject } from "@Services/graph/GraphAxis";
-import { ApiRequest } from "@Services/api/ExtractData";
 import BaseLineGraph from "../base/BaseLineGraph";
+import { fetchAvarageGrade } from "@Services/api/graph/avarageGrade";
 
 const AvarageGradeGraph = () => {
   const [State, SetState] = useState<GraphObject>({ labels: [], data: [] });
 
   useEffect(() => {
     const fetchGraphData = async () => {
-      const [LABELS, DATA] = await new ApiRequest("AvarageGrade").getData();
+      const [LABELS, DATA] = await fetchAvarageGrade();
 
       SetState({
         labels: LABELS,
