+++ src/services/api/graph/totalTimeAtSchool.ts
@@ -1,5 +1,11 @@
 import QS_API from "../WebApi";
 
 export async function fetchTotalTimeSpentAtSchool() {
-  return await QS_API.get("TotalTimeAtSchool");
+  const mockValue = {
+    total: 38.5,
+  };
+
+  return { data: mockValue };
+
+  // return await QS_API.get("TotalTimeAtSchool");
 }
